+++ nodalia-cards.manifest.js
@@ -1,11 +1,11 @@
 export default {
-  "pkgVersion": "2.2.1-alpha.4",
-  "contentSha256_12": "ed3992cb3e49",
+  "pkgVersion": "2.2.1-alpha.5",
+  "contentSha256_12": "d8f03a52a2e1",
   "file": "nodalia-cards.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "editorSha256_12": "fcf4f0088891"
 };
-export const pkgVersion = "2.2.1-alpha.4";
-export const contentSha256_12 = "ed3992cb3e49";
+export const pkgVersion = "2.2.1-alpha.5";
+export const contentSha256_12 = "d8f03a52a2e1";
 export const file = "nodalia-cards.js";